Refactor Dockerfile and app.py, update package-lock.json
This commit refactors the Dockerfile to use a newer version of Node.js (node:18) and updates the app.py file to correctly set the static folder for the Flask app. Additionally, the package-lock.json file is updated to use TypeScript version 4.9.5. The start.sh script is removed as it is no longer needed.